backend/requirements.txt
fastapi>=0.115,<1.0
pydantic>=2.8,<3.0
PyYAML>=6.0,<7.0
uvicorn>=0.30,<1.0